Separate entities from root reducer
